students/periph/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
smbus2==0.5.0
paho-mqtt==2.1.0
